+++ example/app/foundation/foundation-example.component.ts
@@ -3,7 +3,7 @@ import {REACTIVE_FORM_DIRECTIVES, FormGroup, FormControl} from "@angular/forms";
 import {
     DynamicFormService,
     DynamicCheckboxModel,
-    DynamicFormModel
+    DynamicFormControlModel
 } from "@ng2-dynamic-forms/core";
 import {DynamicFormFoundationSitesComponent} from "@ng2-dynamic-forms/ui-foundation";
 import {FOUNDATION_EXAMPLE_MODEL} from "./foundation-example.model";
@@ -21,7 +21,7 @@ import {FOUNDATION_EXAMPLE_MODEL} from "./foundation-example.model";
 
 export class FoundationExampleComponent implements OnInit {
 
-    dynamicFormModel: DynamicFormModel;
+    dynamicFormModel: Array<DynamicFormControlModel>;
     form: FormGroup;
     
     exampleCheckboxControl: FormControl;
@@ -34,10 +34,10 @@ export class FoundationExampleComponent implements OnInit {
 
     ngOnInit() {
 
-        this.form = this.dynamicFormService.createFormGroup(this.dynamicFormModel.group);
+        this.form = this.dynamicFormService.createFormGroup(this.dynamicFormModel);
 
         this.exampleCheckboxControl = <FormControl> this.form.controls["exampleCheckbox"]; // Type assertion for having updateValue method available
-        this.exampleCheckboxModel = <DynamicCheckboxModel> this.dynamicFormModel.findById("exampleCheckbox");
+        this.exampleCheckboxModel = <DynamicCheckboxModel> this.dynamicFormService.findById("exampleCheckbox", this.dynamicFormModel);
         //this.exampleCheckboxControl.valueChanges.subscribe((value: string) => console.log("example checkbox field changed to: ", value, typeof value));
     }
 }