+++ modules/core/index.ts
@@ -1,4 +1,3 @@
-export * from "./src/model/dynamic-form.model";
 export * from "./src/model/dynamic-form-control.model";
 export * from "./src/model/form-group/dynamic-form-group.model";
 export * from "./src/model/form-array/dynamic-form-array.model";