+++ modules/core/src/model/checkbox/dynamic-checkbox.model.ts
@@ -4,6 +4,9 @@ import {getValue} from "../../utils";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX = "CHECKBOX";
 
+export const DYNAMIC_FORM_CONTROL_CHECKBOX_ALIGN_START = "start";
+export const DYNAMIC_FORM_CONTROL_CHECKBOX_ALIGN_END = "end";
+
 export interface CheckboxModelConfig extends DynamicFormValueControlModelConfig {
 
     align?: string;
@@ -15,11 +18,11 @@ export class DynamicCheckboxModel extends DynamicFormValueControlModel<boolean> 
     align: string;
     indeterminate: boolean;
 
-    constructor(config: CheckboxModelConfig = {}, cls?: ClsConfig) {
+    constructor(config: CheckboxModelConfig, cls?: ClsConfig) {
 
         super(config, cls);
 
-        this.align = getValue(config, "align", "start");
+        this.align = getValue(config, "align", DYNAMIC_FORM_CONTROL_CHECKBOX_ALIGN_START);
         this.indeterminate = getValue(config, "indeterminate", false);
         this.type = DYNAMIC_FORM_CONTROL_TYPE_CHECKBOX;