+++ modules/core/src/model/input/dynamic-input.model.spec.ts
@@ -3,119 +3,120 @@ import {AUTOCOMPLETE_ON} from "../../service/dynamic-form-autofill.service";
 
 describe("DynamicInputModel test suite", () => {
 
-    describe("default object test suite", () => {
+    describe("default model test suite", () => {
 
-        let defaultObject: DynamicInputModel;
+        let config = {id: "default"};
+        let defaultModel: DynamicInputModel;
 
         beforeEach(() => {
-            defaultObject = new DynamicInputModel({id: "default"});
+            defaultModel = new DynamicInputModel(config);
         });
         
         it("tests if correct default type property is set", () => {
 
-            expect(defaultObject.type).toBeDefined();
-            expect(defaultObject.type).toEqual(DYNAMIC_FORM_CONTROL_TYPE_INPUT);
+            expect(defaultModel.type).toBeDefined();
+            expect(defaultModel.type).toEqual(DYNAMIC_FORM_CONTROL_TYPE_INPUT);
         });
 
         it("tests if correct default input type property is set", () => {
 
-            expect(defaultObject.inputType).toBeDefined();
-            expect(defaultObject.inputType).toEqual(DYNAMIC_FORM_CONTROL_INPUT_TYPE_TEXT);
+            expect(defaultModel.inputType).toBeDefined();
+            expect(defaultModel.inputType).toEqual(DYNAMIC_FORM_CONTROL_INPUT_TYPE_TEXT);
         });
 
         it("tests if correct default autoComplete property is set", () => {
 
-            expect(defaultObject.autoComplete).toBeDefined();
-            expect(defaultObject.autoComplete).toEqual(AUTOCOMPLETE_ON);
+            expect(defaultModel.autoComplete).toBeDefined();
+            expect(defaultModel.autoComplete).toEqual(AUTOCOMPLETE_ON);
         });
 
         it("tests if correct default autoFocus property is set", () => {
 
-            expect(defaultObject.autoFocus).toBeDefined();
-            expect(defaultObject.autoFocus).toBe(false);
+            expect(defaultModel.autoFocus).toBeDefined();
+            expect(defaultModel.autoFocus).toBe(false);
         });
 
         it("tests if correct default cls properties aree set", () => {
 
-            expect(defaultObject.cls).toBeDefined();
-            expect(defaultObject.cls.element.container).toEqual("");
-            expect(defaultObject.cls.element.control).toEqual("");
-            expect(defaultObject.cls.element.label).toEqual("");
-            expect(defaultObject.cls.grid.container).toEqual("");
-            expect(defaultObject.cls.grid.control).toEqual("");
-            expect(defaultObject.cls.grid.label).toEqual("");
+            expect(defaultModel.cls).toBeDefined();
+            expect(defaultModel.cls.element.container).toEqual("");
+            expect(defaultModel.cls.element.control).toEqual("");
+            expect(defaultModel.cls.element.label).toEqual("");
+            expect(defaultModel.cls.grid.container).toEqual("");
+            expect(defaultModel.cls.grid.control).toEqual("");
+            expect(defaultModel.cls.grid.label).toEqual("");
         });
 
         it("tests if correct default help property is set", () => {
 
-            expect(defaultObject.help).toBeDefined();
-            expect(defaultObject.help).toBeNull();
+            expect(defaultModel.help).toBeDefined();
+            expect(defaultModel.help).toBeNull();
         });
 
         it("tests if correct default label properties aree set", () => {
 
-            expect(defaultObject.label).toBeDefined();
-            expect(defaultObject.label).toBeNull();
+            expect(defaultModel.label).toBeDefined();
+            expect(defaultModel.label).toBeNull();
         });
 
         it("tests if correct default max property is set", () => {
 
-            expect(defaultObject.max).toBeDefined();
-            expect(defaultObject.max).toBeNull();
+            expect(defaultModel.max).toBeDefined();
+            expect(defaultModel.max).toBeNull();
         });
 
         it("tests if correct default maxLength property is set", () => {
 
-            expect(defaultObject.maxLength).toBeDefined();
-            expect(defaultObject.maxLength).toBe(100);
+            expect(defaultModel.maxLength).toBeDefined();
+            expect(defaultModel.maxLength).toBe(100);
         });
         
         it("tests if correct default min property is set", () => {
 
-            expect(defaultObject.min).toBeDefined();
-            expect(defaultObject.min).toBeNull();
+            expect(defaultModel.min).toBeDefined();
+            expect(defaultModel.min).toBeNull();
         });
 
         it("tests if correct default placeholder property is set", () => {
 
-            expect(defaultObject.placeholder).toBeDefined();
-            expect(defaultObject.placeholder).toEqual("");
+            expect(defaultModel.placeholder).toBeDefined();
+            expect(defaultModel.placeholder).toEqual("");
         });
 
         it("tests if correct default readonly property is set", () => {
 
-            expect(defaultObject.readOnly).toBeDefined();
-            expect(defaultObject.readOnly).toBe(false);
+            expect(defaultModel.readOnly).toBeDefined();
+            expect(defaultModel.readOnly).toBe(false);
         });
 
         it("tests if correct default required property is set", () => {
 
-            expect(defaultObject.required).toBeDefined();
-            expect(defaultObject.required).toBe(false);
+            expect(defaultModel.required).toBeDefined();
+            expect(defaultModel.required).toBe(false);
         });
 
         it("tests if correct default spellcheck property is set", () => {
 
-            expect(defaultObject.spellCheck).toBeDefined();
-            expect(defaultObject.spellCheck).toBe(false);
+            expect(defaultModel.spellCheck).toBeDefined();
+            expect(defaultModel.spellCheck).toBe(false);
         });
         
         it("tests if correct default step property is set", () => {
 
-            expect(defaultObject.step).toBeDefined();
-            expect(defaultObject.step).toBeNull();
+            expect(defaultModel.step).toBeDefined();
+            expect(defaultModel.step).toBeNull();
         });
         
         it("tests if correct default prefix property is set", () => {
 
-            expect(defaultObject.prefix).toBeDefined();
-            expect(defaultObject.prefix).toBeNull();
+            expect(defaultModel.prefix).toBeDefined();
+            expect(defaultModel.prefix).toBeNull();
         });
 
         it("tests if correct default suffix property is set", () => {
 
-            expect(defaultObject.suffix).toBeDefined();
-            expect(defaultObject.prefix).toBeNull();
+            expect(defaultModel.suffix).toBeDefined();
+            expect(defaultModel.prefix).toBeNull();
         });
         
     });