+++ modules/core/src/model/input/dynamic-input.model.ts
@@ -43,11 +43,11 @@ export class DynamicInputModel extends DynamicInputControlModel<any> {
 
     private listId: string = null;
 
-    constructor(config: InputModelConfig = {}, cls?: ClsConfig) {
+    constructor(config: InputModelConfig, cls?: ClsConfig) {
 
         super(config, cls);
 
-        this.accept = getValue(config, "accept", "test");
+        this.accept = getValue(config, "accept", null);
         this.inputType = getValue(config, "inputType", DYNAMIC_FORM_CONTROL_INPUT_TYPE_TEXT);
         this.list = getValue(config, "list", null);
         this.max = getValue(config, "max", null);