+++ modules/core/src/model/radio/dynamic-radio-group.model.ts
@@ -1,6 +1,6 @@
 import {ClsConfig} from "../dynamic-form-control.model";
 import {DynamicOptionControlModel, DynamicOptionControlModelConfig} from "../dynamic-option-control.model";
-import {IDynamicFieldset} from "../form-group/dynamic-form-group.model";
+import {DynamicFieldSet} from "../form-group/dynamic-form-group.model";
 import {getValue} from "../../utils";
 
 export const DYNAMIC_FORM_CONTROL_TYPE_RADIO_GROUP = "RADIO_GROUP";
@@ -10,11 +10,11 @@ export interface RadioGroupModelConfig extends DynamicOptionControlModelConfig {
     legend?: string;
 }
 
-export class DynamicRadioGroupModel<T> extends DynamicOptionControlModel<T> implements IDynamicFieldset {
+export class DynamicRadioGroupModel<T> extends DynamicOptionControlModel<T> implements DynamicFieldSet {
 
     legend: string;
 
-    constructor(config: RadioGroupModelConfig = {}, cls?: ClsConfig) {
+    constructor(config: RadioGroupModelConfig, cls?: ClsConfig) {
 
         super(config, cls);