+++ modules/core/src/model/select/dynamic-select.model.ts
@@ -5,7 +5,7 @@ export const DYNAMIC_FORM_CONTROL_TYPE_SELECT = "SELECT";
 
 export class DynamicSelectModel<T> extends DynamicOptionControlModel<T> {
 
-    constructor(config: DynamicOptionControlModelConfig = {}, cls?: ClsConfig) {
+    constructor(config: DynamicOptionControlModelConfig, cls?: ClsConfig) {
 
         super(config, cls);